build: Move tspaths webpack workaround to module
