+++ .gitignore
@@ -12,7 +12,8 @@
 !src/**/*.d.ts
 !examples/**/*.js
 !examples/**/*.d.ts
-!.storybook/middleware.js
+!.storybook/**/*.js
+!.storybook/**/*.d.ts
 
 # Vite
 *.timestamp-*
