+++ .storybook/main.ts
@@ -1,9 +1,6 @@
-import { existsSync } from 'node:fs'
-import { join, parse, resolve } from 'node:path'
-
 import type { StorybookConfig } from '@storybook/react-webpack5'
-import { TsconfigPathsPlugin } from 'tsconfig-paths-webpack-plugin'
-import { NormalModuleReplacementPlugin } from 'webpack'
+
+import webpackTsconfigpaths from './webpack-tsconfigpaths.js'
 
 const config: StorybookConfig = {
   stories: ['../src/**/*.mdx', '../src/**/*.stories.@(js|jsx|ts|tsx)'],
@@ -21,51 +18,8 @@ const config: StorybookConfig = {
   docs: {
     autodocs: 'tag',
   },
-  // UPSTREAM: https://github.com/storybookjs/storybook/issues/14274
-  // Workaround from https://github.com/vercel/next.js/issues/41961#issuecomment-1311091390
   webpackFinal: async (config) => {
-    if (config?.plugins == null) config.plugins = []
-
-    config.plugins.push(
-      new NormalModuleReplacementPlugin(/\.js$/, (resource) => {
-        if (resource.request.startsWith('.')) {
-          const path = resolve(resource.context, resource.request)
-
-          if (!path.includes('node_modules') && !existsSync(path)) {
-            const { dir, name } = parse(path)
-            const extensionlessPath = join(dir, name)
-
-            for (const fallbackExtension of ['.tsx', '.ts', '.js']) {
-              const fallback = `${extensionlessPath}${fallbackExtension}`
-              if (existsSync(fallback)) {
-                resource.request = resource.request.replace(
-                  /\.js$/,
-                  fallbackExtension
-                )
-                break
-              }
-            }
-          }
-        }
-      })
-    )
-
-    config.plugins.push(
-      new NormalModuleReplacementPlugin(/^lib\/.*js$/, (resource) => {
-        resource.request = resource.request.replace(/\.js$/, '')
-      })
-    )
-
-    config.plugins.push(
-      new NormalModuleReplacementPlugin(/^index\.js$/, (resource) => {
-        resource.request = 'src/index.ts'
-      })
-    )
-
-    if (config?.resolve == null) config.resolve = {}
-    if (config?.resolve?.plugins == null) config.resolve.plugins = []
-    config.resolve.plugins.push(new TsconfigPathsPlugin({ baseUrl: '.' }))
-
+    webpackTsconfigpaths(config)
     return config
   },
 }
