+++ .storybook/webpack-tsconfigpaths.d.ts
@@ -1,0 +1,5 @@
+import type { Configuration } from 'webpack'
+
+declare function webpackTsconfigpaths(config: Configuration): void
+
+export default webpackTsconfigpaths
